HEMS: Initial commit
This is a partial implementation of a Home Energy Management System for
a residential solar installation. It enables the user to schedule
appliances in a targeted way, increasing energy self-consumption based
on energy production predictions via weather forecasts.

This application was conceptualized and developed in the context of a
master's thesis in computer science at Freie Universität Berlin,
submitted in June 2021. The thesis manuscript serves as documentation
and is available in this repository in the doc directory.